Move pagination module into separate file
